refactor: rename calendar -> calendars
