+++ frontend/src/app/calendars/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { CalendarLayout } from "@/components/calendar/calendar/CalendarLayout";
+import { CalendarLayout } from "@/components/calendars/calendars/CalendarLayout";
 import { LoginRequired } from "./layout";
 
 export default LoginRequired(CalendarLayout);
