+++ frontend/src/components/calendars/calendars/CalendarLayout/CalendarLayout.tsx
@@ -5,8 +5,8 @@ import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
 import { CalendarSidebar } from '../CalendarSidebar';
 import { RightSidebar } from '../CalendarRightSidebar';
 import { CalendarView } from '../CalendarView';
-import { useCalendar } from '@/lib/calendar/hooks/useCalendar';
-import type { Calendar } from '@/lib/calendar/types/calendarTypes';
+import { useCalendar } from '@/lib/calendars/hooks/useCalendar';
+import type { Calendar } from '@/lib/calendars/types/calendarTypes';
 
 export const CalendarLayout = () => {
   const { calendars, loading, error, createCalendar, updateCalendar, deleteCalendar, fetchCalendars } = useCalendar();
